Add interface to build commit message
For the planner we need to build lots of commit messages, but this
commit adds an interface to make that job easier. Now we only need
to call the commit interface and it will build the message for a
day, but we can easily change that using  `week` or `month` as a
parameters.